medsyn-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pymupdf==1.24.3
google-generativeai==0.5.4
httpx==0.27.0
weasyprint==62.3
pydantic==2.7.1
python-dotenv==1.0.1
sse-starlette==2.1.0
transformers==4.41.0
torch==2.3.0
Pillow==10.3.0
timm==0.9.16
einops==0.7.0
